+++ src/main/kotlin/top/warmthdawn/emss/features/server/Routes.kt
@@ -7,6 +7,8 @@ import org.koin.ktor.ext.inject
 import top.warmthdawn.emss.features.dockerStats.StatsService
 import top.warmthdawn.emss.features.server.dto.ServerInfoDTO
 import top.warmthdawn.emss.utils.R
+import top.warmthdawn.emss.utils.checkPermission
+import top.warmthdawn.emss.utils.checkServerPermission
 
 /**
  *
@@ -24,12 +26,14 @@ fun Route.serverEndpoint() {
         }
 
         post {
+            checkPermission(0)
             val dtoServerInfo = call.receive<ServerInfoDTO>()
             serverService.createServerInfo(dtoServerInfo)
             R.ok()
         }
 
         delete("/{id}"){
+            checkPermission(0)
             val id = call.parameters["id"]!!.toLong()
             serverService.removeServer(id)
             R.ok()
@@ -40,6 +44,7 @@ fun Route.serverEndpoint() {
         }
         post("/{id}") {
             val id = call.parameters["id"]!!.toLong()
+            checkServerPermission(id, 3)
             val dto = call.receive<ServerInfoDTO>()
             serverService.updateServerInfo(id, dto)
             R.ok()
@@ -50,21 +55,25 @@ fun Route.serverEndpoint() {
         }
         post("/{id}/start") {
             val id = call.parameters["id"]!!.toLong()
+            checkServerPermission(id, 3)
             serverService.start(id)
             R.ok()
         }
         post("/{id}/stop") {
             val id = call.parameters["id"]!!.toLong()
+            checkServerPermission(id, 3)
             serverService.stop(id)
             R.ok()
         }
         post("/{id}/restart") {
             val id = call.parameters["id"]!!.toLong()
+            checkServerPermission(id, 3)
             serverService.restart(id)
             R.ok()
         }
         post("/{id}/terminate") {
             val id = call.parameters["id"]!!.toLong()
+            checkServerPermission(id, 3)
             serverService.terminate(id)
             R.ok()
         }
